Finish Place Block Implementation and Adjust Game Struct Fields
Finished implementing the place block function, it nows checks for rows to clear and call the clear row function on them.

Adjust the game struct to map row number to the blocks held in that row. This makes it real easy to get all of the blocks in a certain row which will be very useful for clear_row() and makes checking for a row to clear trivial.

Additionally, number of lines cleared will now be tracked to be used later for scoring and leveling up.